requirements.txt
prometheus_client
RPi.GPIO